requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
langchain>=1.2.0
langchain-core>=1.2.0
langchain-ollama>=1.0.0
langchain-openai>=0.3.0
pydantic>=2.10.0
python-dotenv==1.0.1
tuya-connector-python>=0.1.2
python-telegram-bot>=20.0
langgraph>=1.1.0
pytest>=8.0.0
